fs/dyn_sync_cntrl: dynamic sync control 1.6 for OPO
While screen is on, file sync is disabled.
When screen is off, a file sync is called to flush all outstanding writes
and restore file sync operation as normal.

The dynamic sync control interface uses a hook in the OPO backlight driver.

By andip71 (aka Lord Boeffla)

Credits for original implementation and concept to Faux123

Signed-off-by: randomlyjobless&bored <randomlybored@openmailbox.org>

Conflicts:
	drivers/video/backlight/lm3630_bl.c